Add subst functor for regex substitution
Syntax is subst(regex, replacement, text). It replaces all matches of
regex in text by the replacement expanding backreferences using
std::regex_replace. Global substitution is used because it can easily be
extended to replace only the first match by tacking a (.*) on the end of
the pattern and including the subexpression in the replacement.